association DataManager and DefaultTask
