+++ ToDoApp/Extension/DataManager.swift
@@ -24,4 +24,15 @@ class DataManager {
             Task(imageName: "Vacation", name: nil, completed: false),
             Task(imageName: "Work", name: nil, completed: false)]
     }
+    
+    class func createTask() -> [Task] {
+        
+        return [Task(imageName: "Family", name: "Заказать путевку", completed: false),
+                Task(imageName: "Grocery", name: "Купить молоко", completed: false),
+                Task(imageName: "Party", name: "Купить хлопушки", completed: false),
+                Task(imageName: "Shopping", name: "Купить сметану", completed: true),
+                Task( imageName: "Baby", name: "Купить памперсы =)", completed: false),
+                Task(imageName: "Vacation", name: "Купить мяч", completed: false),
+                Task(imageName: "Work", name: "Купить SSD", completed: false)]
+    }
 }
